lambda/requirements.txt
# requests and PIL handler
requests 
Pillow